added seperate styles file for quick access
